Separate Forge implementation
