+++ controlnet_ext/__init__.py
@@ -1,4 +1,17 @@
-from .controlnet_ext import ControlNetExt, controlnet_exists, controlnet_forge, get_cn_models
+try:
+    from .controlnet_ext_forge import (
+        ControlNetExt,
+        controlnet_exists,
+        controlnet_forge,
+        get_cn_models,
+    )
+except ImportError:
+    from .controlnet_ext import (
+        ControlNetExt,
+        controlnet_exists,
+        controlnet_forge,
+        get_cn_models,
+    )
 
 __all__ = [
     "ControlNetExt",
